recommend runtime policy using karmor
- Added the ability to query discovery engine to get summary details
- Added the ability to create dynamic policy based on k8s service account access data from discovery engine

To-Do
- Optimise the functions
- Show details of generated policies on report file
- Create single policy if path/dir has multiple `fromSource` value

Addresses https://github.com/kubearmor/kubearmor-client/issues/112#issuecomment-1216476912

+++ recommend/recommend.go
@@ -74,7 +74,7 @@ func finalReport() {
 // Recommend handler for karmor cli tool
 func Recommend(c *k8s.Client, o Options) error {
 	var err error
-
+	options = o
 	if err = createOutDir(o.OutDir); err != nil {
 		return err
 	}
@@ -85,21 +85,31 @@ func Recommend(c *k8s.Client, o Options) error {
 
 	o.Images = unique(o.Images)
 	o.Tags = unique(o.Tags)
-	options = o
-	for _, img := range o.Images {
-		tempDir, err = os.MkdirTemp("", "karmor")
-		if err != nil {
-			log.WithError(err).Fatal("could not create temp dir")
+	if len(o.Images) > 1 {
+
+		for _, img := range o.Images {
+			tempDir, err = os.MkdirTemp("", "karmor")
+			if err != nil {
+				log.WithError(err).Fatal("could not create temp dir")
+			}
+			err = imageHandler(img)
+			if err != nil {
+				log.WithError(err).WithFields(log.Fields{
+					"image": img,
+				}).Error("could not handle container image")
+			}
+			_ = os.RemoveAll(tempDir) // rm -rf tempDir
 		}
-		err = imageHandler(img)
+
+	} else {
+		err = RuntimePolicy(&o)
 		if err != nil {
 			log.WithError(err).WithFields(log.Fields{
-				"image": img,
-			}).Error("could not handle container image")
+				"labels":    o.UseLabels,
+				"namespace": o.UseNamespace,
+			}).Error("could not create runtime policy")
 		}
-		_ = os.RemoveAll(tempDir) // rm -rf tempDir
 	}
-
 	finalReport()
 	return nil
 }
